# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API for the AI Music Production Studio
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /storage/uploads/request-url:
    post:
      tags: [sources]
      operationId: requestSourceUploadUrl
      summary: Request a protected direct-upload URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadUrlRequest"
      responses:
        "200":
          description: Direct upload target
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadUrlResponse"
        "401":
          description: Login required
...
components:
...
  schemas:
...
    UploadUrlRequest:
      type: object
      required: [name, size, contentType]
      properties:
        name: { type: string, minLength: 1 }
        size: { type: number, minimum: 1, maximum: 524288000 }
        contentType: { type: string, minLength: 1 }
    UploadUrlResponse:
      type: object
      required: [uploadURL, objectPath, metadata]
      properties:
        uploadURL: { type: string }
        objectPath: { type: string }
        metadata:
          $ref: "#/components/schemas/UploadUrlRequest"
...